Inital work on allowing disabling of default actions for a Resource

+++ lib/active_admin/resource/action_items.rb
@@ -9,7 +9,6 @@ module ActiveAdmin
       # initialized
       def initialize(*args)
         super
-        add_default_action_items
       end
 
       # @return [Array] The set of action items for this resource
@@ -34,6 +33,11 @@ module ActiveAdmin
       #
       # @return [Array] Array of ActionItems for the controller actions
       def action_items_for(action)
+        unless @default_actions_added
+          add_default_action_items
+          @default_actions_added = true
+        end
+
         action_items.select{|item| item.display_on?(action) }
       end
 
@@ -41,28 +45,45 @@ module ActiveAdmin
       def clear_action_items!
         @action_items = []
       end
+      
+      # Allows disabling of the default actions (new / edit / destroy)
+      # based on the conditions provided
+      #     :if
+      #     :unless
+      def disable_action_item_for(type, options)
+        @disabled_default_actions ||= {}
+        @disabled_default_actions[type.to_s] = options
+      end
 
+      def default_action_disabled?(type, resource = nil)
+        return false unless @disabled_default_actions
+        return false unless @disabled_default_actions[type]
+        return @disabled_default_actions[type][:if].call(resource) if @disabled_default_actions[type][:if]
+        return !@disabled_default_actions[type][:unless].call(resource) if @disabled_default_actions[type][:unless]
+        return false
+      end
+      
       private
 
       # Adds the default action items to each resource
       def add_default_action_items
         # New Link on all actions except :new and :show
         add_action_item :except => [:new, :show] do
-          if controller.action_methods.include?('new')
+          if controller.action_methods.include?('new') && !active_admin_config.default_action_disabled?('new')
             link_to(I18n.t('active_admin.new_model', :model => active_admin_config.resource_name), new_resource_path)
           end
-        end
+        end 
 
         # Edit link on show
         add_action_item :only => :show do
-          if controller.action_methods.include?('edit')
+          if controller.action_methods.include?('edit') && !active_admin_config.default_action_disabled?('edit', resource)
             link_to(I18n.t('active_admin.edit_model', :model => active_admin_config.resource_name), edit_resource_path(resource))
           end
         end
 
         # Destroy link on show
         add_action_item :only => :show do
-          if controller.action_methods.include?("destroy")
+          if controller.action_methods.include?("destroy") && !active_admin_config.default_action_disabled?('destroy', resource)
             link_to(I18n.t('active_admin.delete_model', :model => active_admin_config.resource_name),
               resource_path(resource),
               :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'))
